requirements.txt
absl-py==0.13.0
asgiref==3.4.1
astunparse==1.6.3
beautifulsoup4==4.6.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
colorthief==0.2.1
cycler==0.10.0
deepface==0.0.68
easydict==1.9
emoji==1.6.1
fastapi==0.68.1
email-validator
filelock==3.0.12
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gdown==3.13.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h11==0.12.0
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
JPype1==1.3.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
konlpy==0.5.2
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
Morfessor==2.0.6
mtcnn==0.1.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.3.2
Pillow==8.3.2
polyglot==16.7.4
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
pydantic==1.8.2
PyICU==2.7.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytrends==4.7.3
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retina-face==0.0.5
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
starlette==0.14.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tqdm==4.62.2
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.26.6
uvicorn==0.15.0
Werkzeug==2.0.1
wrapt
xgboost
###termcolor==1.1.0 depend on torch
###threadpoolctl==2.2.0
###torch
###torchvision==0.10.0

